Enabling MigrationServices to run via resque
