+++ lib/curate/migration_services/migration_containers.rb
@@ -5,6 +5,7 @@ module Curate
   end
 end
 
+require 'curate/migration_services/base_migrator'
 Dir[File.expand_path('../migration_containers/**/*.rb', __FILE__)].each do |filename|
-  require 'filename'
+  require filename
 end
