--- lib/curate/migration_services/unsaved_digital_object.rb
@@ -1,11 +1,0 @@
-module Curate
-  module MigrationServices
-    class UnsavedDigitalObject < ActiveFedora::UnsavedDigitalObject
-      attr_reader :repository
-      def initialize(repository, *args, &block)
-        @repository = repository
-        super(*args, &block)
-      end
-    end
-  end
-end
